feat: Complete SwiftUI Core Showcase implementation with comprehensive demos
- Implemented all 29 SwiftUI demos across 10 categories
- Added comprehensive keyframe animation, spring animation, and transition demos
- Implemented complex gestures, visual effects, and custom layout demos
- Added accessibility, navigation, and interactive control demos
- Fixed build issues and compilation errors
- Added app icons and visual assets
- Achieved 100% demo coverage with advanced SwiftUI features including:
  - Keyframe animations with interactive controls
  - Custom layout protocols and flow layouts
  - Advanced gesture recognizers and multi-touch support
  - Visual effects with materials and blur effects
  - Mask and clipping operations with custom shapes
  - Glass effects and morphing animations
  - Comprehensive progress indicators and controls
  - State management and data binding patterns

+++ swiftuicoreshowcase/Views/Content/ShapeDemo.swift
@@ -356,12 +356,12 @@ struct StarShape: Shape {
             let midAngle = (angle + nextAngle) / 2
             
             let outerPoint = CGPoint(
-                x: center.x + cos(angle) * radius,
-                y: center.y + sin(angle) * radius
+                x: center.x + CGFloat(cos(angle)) * radius,
+                y: center.y + CGFloat(sin(angle)) * radius
             )
             let innerPoint = CGPoint(
-                x: center.x + cos(midAngle) * innerRadius,
-                y: center.y + sin(midAngle) * innerRadius
+                x: center.x + CGFloat(cos(midAngle)) * innerRadius,
+                y: center.y + CGFloat(sin(midAngle)) * innerRadius
             )
             
             if i == 0 {
@@ -409,8 +409,8 @@ struct HexagonShape: Shape {
         for i in 0..<6 {
             let angle = Double(i) * .pi / 3
             let point = CGPoint(
-                x: center.x + cos(angle) * radius,
-                y: center.y + sin(angle) * radius
+                x: center.x + CGFloat(cos(angle)) * radius,
+                y: center.y + CGFloat(sin(angle)) * radius
             )
             
             if i == 0 {
